Move preprocessing to seperate phase
