+++ effekt/shared/src/main/scala/effekt/generator/llvm/LLVM.scala
@@ -59,7 +59,7 @@ class LLVM extends Compiler[String] {
     // 1. Split off Deadcode from Optimizer
     // 2. Do Show (hope it still runs)
     // 3. Run Optimizer
-    val afterCore = allToCore(Core) andThen Aggregate andThen optimizer.Deadcode andThen core.Show andThen core.Mono andThen optimizer.Optimizer
+    val afterCore = allToCore(Core) andThen Aggregate andThen optimizer.Deadcode andThen core.Show andThen core.Preprocess andThen core.Mono andThen optimizer.Optimizer
     val afterMachine = afterCore andThen Machine map { case (mod, main, prog) => prog }
     val afterLLVM = afterMachine map {
       case machine.Program(decls, defns, entry) =>
